Incremental
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/contrib@2308 67a373e5-eb02-0410-a15c-ee090a768436

+++ fcptools/ezFCPlib/ezFCPlib.h
@@ -376,7 +376,7 @@ typedef struct {
 extern char  _fcpID[4];
 
 extern char *_fcpHost;
-extern int   _fcpPort;
+extern unsigned short _fcpPort;
 extern int   _fcpHtl;
 extern int   _fcpRawmode;
